Add the ExportEngine trait and implement the mbox export engine
This commit adds the ExportEngine trait and implements an MboxEngine for
exporting backed-up emails to mbox format. It also adds the ExportRunner
and hooks it into a newly exposed `export` subcommand.

Some caveats of the current
implementation is that it is not based on streaming (ie iterators) and
all emails need to be loaded into memory before being passed on to the
engine.

Current mbox engine has been tested with python3 mailbox module which is
seems happy to load and parse a full export of my Gmail account.